+++ src/xfp/xresult/xtry.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 from typing import Never, ParamSpec, TypeVar
 from xfp.xresult._xresult import Xresult, XRBranch
 from xfp.xresult.xeither import Xeither
-from xfp.types import F0, F1
+from xfp.functions import F0, F1
 
 P = ParamSpec("P")
 X = TypeVar("X")
